requirements.txt
aiogram==3.5.0
python-dotenv==1.0.1
openai==0.28.1
pytesseract==0.3.10
Pillow==10.3.0
fastapi==0.111.0
uvicorn==0.29.0
fpdf==1.7.2
aiofiles==23.2.1
aiohttp==3.9.5
